+++ modules/graphics/src/include/GOFigure.hpp
@@ -26,9 +26,7 @@ private:
     GOWindow* m_win;
 
     void
-    refreshInnerPositionProperty();
-    void
-    refreshOuterPositionProperty();
+    refreshDrawLaterProperty();
 
 public:
     GOFigure(GOWindow* win, int number);