+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- `drawnow`: Update figures and process callbacks
+- `DrawLater` property added to `figure` graphics object.
 - `interp1` linear interpolation 1D.
 - [736](http://github.com/Nelson-numerical-software/nelson/issues/736): `bone`, `cool`, `copper`, `hot`, `jet`, `pink`, `turbo`, `viridis`, `white` colormaps.
 - `CHANGELOG` 0.7.x family.